Save connection details as GlobalSettings.
Server URL, Server Port, and SlotName will be saved as GlobalSettings.

Also adjusts OnLoadLocal to noop if SlotName is empty.  For a real save,
this will never be the case -- but apparently OnLoadLocal is also called
for the menu and in other cases which was causing the restored global
settings to be consistently overridden with their defaults.