Adds support for generic location costs.
AP can now include location_costs in slotdata, of the form
```{location: {currency: cost, currency: cost...}}```

If location_costs are present, previous placement handler logic is ignored
(it's only kept for backwards compatibility) and the location costs is
the sole determining factor of all costs.

This fixes vanilla costs (stags, maps, fountain), as well as enabling the
possibility of... weird cost randomization options going in the future
like stag stations requiring grubs and what not.

Oh, also, we reverse the location list before adding it, as a quick and
dirty fix to get things sorted in the right order.