Revamped DeathLink support to be more PlayMakerFSM-based.
This now directly alters the Death animation FSM, which should prove more
reliable in general though slightly more fragile when other mods are
involved if they alter the same FSM.

Notably gained by this is the ability to completely bypass creating a
shade.  As a result, it is no longer neccessary to snapshot shades or
ensure that they are snapshotted or not at the correct time.