Contains assorted placement bugfixes.
IC's Finder.GetLocation() returns a CLONE of the location, which makes it
unsuitable for use as a dictionary key -- subsequent calls will produce
a value that does not exist in the dictionary.

As a result, our caching of previously created placements was not caching
anything at all.  This is incidentally why placements were happening in
reverse order.

Also, all of our default shop items/destroyed rewards were only working if
the shop had zero placements, as the first item placed would destroy that
placement and effectively overwrite it.